fix(orders-table): approve USDT
