+++ apps/cowswap-frontend/src/common/hooks/useApproveCurrency.ts
@@ -1,0 +1,22 @@
+import { useCallback } from 'react'
+
+import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
+
+import { useShouldZeroApprove, useZeroApprove } from 'modules/zeroApproval'
+
+import { useTradeApproveCallback } from '../containers/TradeApprove'
+
+// TODO: allocate a module for all approval process logic
+export function useApproveCurrency(amountToApprove: CurrencyAmount<Currency> | undefined) {
+  const tradeApproveCallback = useTradeApproveCallback(amountToApprove)
+  const shouldZeroApprove = useShouldZeroApprove(amountToApprove)
+  const zeroApprove = useZeroApprove(amountToApprove?.currency)
+
+  return useCallback(async () => {
+    if (shouldZeroApprove) {
+      await zeroApprove()
+    }
+
+    await tradeApproveCallback()
+  }, [tradeApproveCallback, zeroApprove, shouldZeroApprove])
+}
